+++ src/index.ts
@@ -1,0 +1,6 @@
+export * from './constants';
+export * from './decoded-id-token';
+export * from './firebase-auth.config';
+export * from './firebase.strategy';
+export * from './firebase-auth.module';
+export * from './firebase-auth.guard';