+++ CHANGELOG.md
@@ -1,1 +1,7 @@
-RELEASE 1.0.0
+# Changelog
+
+## [1.0.0] - 2021-03-17
+### Initial Release
+- Add custom passport strategy named FirebaseStrategy, extending
+  the PassportJwt Strategy.
+- Create a dynamic module for initialization of FirebaseStrategy